pyproject.toml
[tool.poetry]
name = "elms"
version = "0.1.0"
description = "An Enhanced E-Learning Management System (E-LMS)"
authors = ["Abdelrahman Farah Mohammed <abdelrahman.mohammed@bld.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.7"
djangorestframework = "^3.14.0"
djoser = "^2.1.0"
social-auth-core = "^4.3.0"
django-cors-headers = "^3.13.0"
mysqlclient = "^2.1.1"
Pillow = "^9.4.0"
drf-nested-routers = "^0.93.4"
google-auth = "^2.17.2"
google-api-python-client = "^2.84.0"
google-auth-oauthlib = "^1.0.0"
oauth2client = "^4.1.3"


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^3.8.1"
autopep8 = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
